+++ leather/chart.py
@@ -13,7 +13,7 @@ from leather.series import Series, CategorySeries
 from leather.shapes import Bars, Columns, Dots, Line, GroupedBars
 import leather.svg as svg
 from leather import theme
-from leather.utils import X, Y, Z, Box, IPythonSVG
+from leather.utils import X, Y, Z, DIMENSION_NAMES, Box, IPythonSVG, warn
 
 
 class Chart(object):
@@ -188,7 +188,13 @@ class Chart(object):
         if not scale:
             scale = Scale.infer(self._layers, dimension, self._types[dimension])
         else:
-            scale = scale
+            for series, shape in self._layers:
+                if not scale.contains(series.min(dimension)) or not scale.contains(series.max(dimension)):
+                    d = DIMENSION_NAMES[dimension]
+                    warn('Data contains values outside %s scale domain. All data points may not be visible on the chart.' % d)
+
+                    # Only display once per axis
+                    break
 
         if not axis:
             axis = Axis()