+++ leather/utils.py
@@ -6,6 +6,9 @@ from decimal import Decimal
 import math
 import six
 import sys
+import warnings
+
+import six
 
 try:
     __IPYTHON__
@@ -28,6 +31,8 @@ Y = 1
 Z = 2
 
 
+DIMENSION_NAMES = ['X', 'Y', 'Z']
+
 #: Data structure for representing margins or other CSS-edge like properties
 Box = namedtuple('Box', ['top', 'right', 'bottom', 'left'])
 
@@ -36,6 +41,15 @@ Datum = namedtuple('Datum', ['i', 'x', 'y', 'z', 'row'])
 
 #: Dummy object used in place of a series when rendering legends for categories
 DummySeries = namedtuple('DummySeries', ['name'])
+
+
+formatwarning_orig = warnings.formatwarning
+warnings.formatwarning = lambda message, category, filename, lineno, line=None: \
+    formatwarning_orig(message, category, filename, lineno, line='')
+
+warn = warnings.warn
+warnings.resetwarnings()
+warnings.simplefilter('always')
 
 
 # In Python 3.5 use builtin C implementation of `isclose`