+++ tests/test_scales.py
@@ -33,6 +33,12 @@ class TestLinear(leather.LeatherTestCase):
         with self.assertRaises(ValueError):
             leather.Linear(10, 0)
 
+    def test_no_spread(self):
+        scale = leather.Linear(0, 0)
+
+        self.assertEqual(scale.project(0, 0, 10), 0)
+        self.assertEqual(scale.project(1, 0, 10), 10)
+
     def test_ticks(self):
         scale = leather.Linear(0, 10)
 
@@ -47,6 +53,16 @@ class TestLinear(leather.LeatherTestCase):
         self.assertEqual(scale.project(Decimal(5), Decimal(10), Decimal(41)), Decimal(25.5))
 
         self.assertEqual(scale.ticks()[1], Decimal(2.5))
+
+    def test_contains(self):
+        scale = leather.Linear(-5, 5)
+
+        self.assertTrue(scale.contains(-5))
+        self.assertTrue(scale.contains(0))
+        self.assertTrue(scale.contains(5))
+        self.assertFalse(scale.contains(-6))
+        self.assertFalse(scale.contains(6))
+
 
 class TestOrdinal(leather.LeatherTestCase):
     def test_project(self):
@@ -71,6 +87,14 @@ class TestOrdinal(leather.LeatherTestCase):
         scale = leather.Ordinal(['a', 'b', 'c', 'd'])
 
         self.assertEqual(scale.ticks(), ['a', 'b', 'c', 'd'])
+
+    def test_contains(self):
+        scale = leather.Ordinal(['a', 'b', 'c', 'd'])
+
+        self.assertTrue(scale.contains('a'))
+        self.assertFalse(scale.contains('aa'))
+        self.assertFalse(scale.contains('e'))
+        self.assertFalse(scale.contains(5))
 
 
 class TestTemporal(leather.LeatherTestCase):
@@ -104,88 +128,11 @@ class TestTemporal(leather.LeatherTestCase):
         self.assertEqual(ticks[0], date(2010, 1, 1))
         self.assertEqual(ticks[-1], date(2014, 1, 1))
 
+    def test_contains(self):
+        scale = leather.Temporal(date(2010, 1, 1), date(2014, 1, 1))
 
-class TestYears(leather.LeatherTestCase):
-    def test_project(self):
-        scale = leather.Years(date(2010, 1, 1), date(2014, 1, 1))
-
-        self.assertEqual(scale.project(date(2011, 1, 1), 0, 20), 6)
-        self.assertEqual(scale.project(date(2012, 1, 1), 0, 20), 10)
-        self.assertEqual(scale.project(date(2009, 1, 1), 0, 20), -6)
-
-        scale = leather.Years(datetime(2010, 1, 1), datetime(2014, 1, 1))
-
-        self.assertEqual(scale.project(datetime(2011, 1, 1), 0, 20), 6)
-        self.assertEqual(scale.project(datetime(2012, 1, 1), 0, 20), 10)
-        self.assertEqual(scale.project(datetime(2009, 1, 1), 0, 20), -6)
-
-        scale = leather.Years(2010, 2014)
-
-        self.assertEqual(scale.project(2011, 0, 20), 6)
-        self.assertEqual(scale.project(2012, 0, 20), 10)
-        self.assertEqual(scale.project(2009, 0, 20), -6)
-
-    def test_project_interval(self):
-        scale = leather.Years(date(2010, 1, 1), date(2014, 1, 1))
-
-        self.assertEqual(scale.project_interval(date(2011, 1, 1), 0, 20), (4.2, 7.8))
-
-    def test_ticks(self):
-        scale = leather.Years(date(2010, 1, 1), date(2014, 1, 1))
-
-        self.assertEqual(scale.ticks(), [
-            date(2010, 1, 1),
-            date(2011, 1, 1),
-            date(2012, 1, 1),
-            date(2013, 1, 1),
-            date(2014, 1, 1)
-        ])
-
-
-class TestMonths(leather.LeatherTestCase):
-    """
-    See notes for :class:`.TestTemporal`.
-    """
-    def test_project(self):
-        scale = leather.Months(date(2010, 1, 1), date(2014, 1, 1))
-
-        self.assertAlmostEqual(scale.project(date(2011, 1, 1), 0, 48), 12, 0)
-        self.assertAlmostEqual(scale.project(date(2012, 1, 1), 0, 48), 24, 0)
-        self.assertAlmostEqual(scale.project(date(2008, 12, 1), 0, 48), -12, 0)
-
-        scale = leather.Months(datetime(2010, 1, 1), datetime(2014, 1, 1))
-
-        self.assertAlmostEqual(scale.project(datetime(2011, 1, 1), 0, 48), 12, 0)
-        self.assertAlmostEqual(scale.project(datetime(2012, 1, 1), 0, 48), 24, 0)
-        self.assertAlmostEqual(scale.project(datetime(2008, 12, 1), 0, 48), -12, 0)
-
-        with self.assertRaises(ValueError):
-            scale = leather.Months(2010, 2014)
-
-    def test_project_interval(self):
-        scale = leather.Months(date(2010, 1, 1), date(2014, 1, 1))
-
-        a, b = scale.project_interval(date(2011, 1, 1), 0, 48)
-        self.assertAlmostEqual(a, 11.5, 0)
-        self.assertAlmostEqual(b, 12.5, 0)
-
-    def test_ticks(self):
-        scale = leather.Months(date(2010, 1, 1), date(2014, 1, 1))
-
-        self.assertEqual(scale.ticks(), [
-            date(2010, 1, 1),
-            date(2011, 1, 1),
-            date(2012, 1, 1),
-            date(2013, 1, 1),
-            date(2014, 1, 1)
-        ])
-
-        scale = leather.Months(date(2010, 1, 1), date(2012, 1, 1))
-
-        self.assertEqual(scale.ticks(), [
-            date(2010, 1, 1),
-            date(2010, 7, 1),
-            date(2011, 1, 1),
-            date(2011, 7, 1),
-            date(2012, 1, 1)
-        ])
+        self.assertTrue(scale.contains(date(2010, 1, 1)))
+        self.assertTrue(scale.contains(date(2012, 6, 3)))
+        self.assertTrue(scale.contains(date(2014, 1, 1)))
+        self.assertFalse(scale.contains(date(2009, 12, 31)))
+        self.assertFalse(scale.contains(date(2014, 1, 2)))